Finished up part 3
I went back and double checked part 3 and followed each and every instruction. I fixed up the bins so that it's now based on max and min of the data and the predetermined width of the bins. If anyone feels that this might be unreasonable or just doesn't make sense, feel free to change it. I found the intrinsic scatter based on my understanding of scatters.